the visibility of courses are now limited by role, instructors only see courses they are part of, and admins can see all courses. tests from my old branch are now added and all tests are in their own directory.
