+++ ta_scheduler/test/acceptance_tests/test-acceptance-test-course-creation.py
